requirements.txt
annotated-types==0.7.0
anyio==4.8.0
archspec
asttokens
black==25.1.0
boltons
brotli
certifi
cffi
charset-normalizer
click==8.1.8
colorama
comm
cryptography
debugpy
decorator
distro
exceptiongroup
executing
frozendict
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna
importlib-metadata
ipykernel
ipython
ipython-pygments-lexers
isort==6.0.1
jedi
jiter==0.8.2
jsonpatch
jsonpointer==2.1
jupyter-client
jupyter-core
keyboard==0.13.5
matplotlib-inline
mypy-extensions==1.0.0
nest-asyncio
openai==1.65.4
packaging
parso
pathspec==0.12.1
pickleshare
platformdirs
pluggy
prompt-toolkit
psutil
pure-eval
pycosat
pycparser
pydantic==2.10.6
pydantic-core==2.27.2
pyfirmata2==2.5.1
pygments
pyserial==3.5
pysocks
python-dateutil
pyzmq
requests
ruamel-yaml
ruamel-yaml-clib
setuptools==75.1.0
six
sniffio==1.3.1
stack-data
tornado
tqdm
traitlets
truststore
typing-extensions
urllib3
wcwidth
wheel==0.44.0
win-inet-pton
zipp
zstandard